+++ src/Encamina.Enmarcha.SemanticKernel/MemoryManager.cs
@@ -5,6 +5,7 @@ using System.Runtime.CompilerServices;
 using System.Text.Json;
 
 using Encamina.Enmarcha.SemanticKernel.Abstractions;
+using Encamina.Enmarcha.SemanticKernel.Abstractions.Events;
 
 using Microsoft.Extensions.Logging;
 using Microsoft.SemanticKernel;
@@ -22,14 +23,25 @@ public class MemoryManager : IMemoryManager
 
     private readonly ILogger logger;
 
+    /// <inheritdoc/>
+    public event EventHandler<MemoryStorageEventArgs> MemoryStorageEvent;
+
     /// <summary>
     /// Initializes a new instance of the <see cref="MemoryManager"/> class.
     /// </summary>
     /// <param name="memoryStore">A valid instance of a <see cref="IMemoryStore"/> to manage.</param>
-    public MemoryManager(IMemoryStore memoryStore, ILogger<MemoryManager> logger)
+    /// <param name="memoryStorageEventHandlerDelegate">A delegate to handle events from <see cref="MemoryStorageEvent"/>.</param>
+    /// <param name="logger">Log service.</param>
+    public MemoryManager(IMemoryStore memoryStore, Action<string, MemoryStorageEventArgs> memoryStorageEventHandlerDelegate, ILogger<MemoryManager> logger)
     {
-        this.logger = logger;
         MemoryStore = memoryStore;
+
+        if (memoryStorageEventHandlerDelegate is not null)
+        {
+            MemoryStorageEvent += (sender, args) => memoryStorageEventHandlerDelegate?.Invoke(sender?.ToString() ?? MemoryStorageEventEnum.Undefined.GetEnumDescription(), args);
+        }
+
+        this.logger = logger;
     }
 
     /// <inheritdoc/>
@@ -46,6 +58,10 @@ public class MemoryManager : IMemoryManager
         }
 
         await SaveChunks(memoryId, collectionName, chunks, metadata, kernel, cancellationToken);
+
+        MemoryStorageEvent?.Invoke(
+            MemoryStorageEventEnum.Upsert.GetEnumDescription(),
+            new() { EventType = MemoryStorageEventEnum.Upsert, MemoryId = memoryId, CollectionName = collectionName });
     }
 
     /// <inheritdoc/>
@@ -56,6 +72,10 @@ public class MemoryManager : IMemoryManager
         if (chunkSize > 0)
         {
             await DeleteMemoryAsync(memoryId, collectionName, chunkSize, cancellationToken);
+
+            MemoryStorageEvent?.Invoke(
+                MemoryStorageEventEnum.Delete.GetEnumDescription(),
+                new() { EventType = MemoryStorageEventEnum.Delete, MemoryId = memoryId, CollectionName = collectionName });
         }
     }
 
@@ -71,6 +91,10 @@ public class MemoryManager : IMemoryManager
 
         var memoryRecords = await MemoryStore.GetBatchAsync(collectionName, Enumerable.Range(0, chunkSize).Select(i => BuildMemoryIdentifier(memoryId, i)), cancellationToken: cancellationToken)
                                              .ToListAsync(cancellationToken);
+
+        MemoryStorageEvent?.Invoke(
+            MemoryStorageEventEnum.Get.GetEnumDescription(),
+            new() { EventType = MemoryStorageEventEnum.Get, MemoryId = memoryId, CollectionName = collectionName });
 
         return new MemoryContent
         {
@@ -107,7 +131,12 @@ public class MemoryManager : IMemoryManager
 
         await foreach (var item in memoryRecordsUniqueIdentifiers)
         {
+            MemoryStorageEvent?.Invoke(
+                MemoryStorageEventEnum.UpsertBatch.GetEnumDescription(),
+                new() { EventType = MemoryStorageEventEnum.UpsertBatch, MemoryId = item, CollectionName = collectionName });
+
             logger.LogInformation($@"Processed memory record {item}.");
+
             yield return item;
         }
     }